% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_initialization.R
\name{add_arc_DBN}
\alias{add_arc_DBN}
\title{Function for arc addiction in Dynamic Bayesian Network}
\usage{
add_arc_DBN(DBN, from, to, cycle_OK = TRUE)
}
\arguments{
\item{DBN}{object of class 'DBN'}

\item{from}{two-element character vector c(var_name, time), with time is \eqn{t_0}, \eqn{t} or \eqn{t-k} (with \eqn{0 <= k <= markovian order})}

\item{to}{two-element character vector c(var_name, time), with time is \eqn{t_0}, \eqn{t} or \eqn{t-k} (with \eqn{0 <= k <= markovian order})}

\item{cycle_OK}{boolean value, TRUE if cycle are admitted (default)}
}
\value{
object of class 'DBN'
}
\description{
Function for arc addiction in Dynamic Bayesian Network
}
\examples{
DBN_example <- add_arc_DBN(DBN=DBN_example,from=c('A','t_0'),to=c('R','t_0'))
DBN_example <- add_arc_DBN(DBN=DBN_example,from=c('S','t'),to=c('O','t'))
DBN_example <- add_arc_DBN(DBN=DBN_example,from=c('S','t-1'),to=c('S','t'))
}
