% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_structure_learning.R
\name{blacklist.DBN}
\alias{blacklist.DBN}
\title{It returns a blacklist matrix represented by the pair of temporal nodes which are forbidden}
\usage{
blacklist.DBN(static_nodes = c(), dynamic_nodes, markov_order = 1)
}
\arguments{
\item{static_nodes}{character vector of time independent variables}

\item{dynamic_nodes}{character vector of time dependent variables}

\item{markov_order}{integer}
}
\value{
matrix of node pairs which edge is forbidden
}
\description{
It returns a blacklist matrix represented by the pair of temporal nodes which are forbidden
}
\examples{
blacklist.DBN(dynamic_nodes = ('A','B'))
}
