% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_structure_learning.R
\name{blacklist.temp.DBN}
\alias{blacklist.temp.DBN}
\title{Given a data.frame it returns a blacklist matrix represented by the pair of temporal nodes which are forbidden}
\usage{
blacklist.temp.DBN(data)
}
\arguments{
\item{data}{a data.frame with column 'Time', 'Sample_id' and the other columns with the values of the time series of 'Sample_id' at time 'Time'}
}
\value{
matrix of node pairs which edge is forbidden
}
\description{
Given a data.frame it returns a blacklist matrix represented by the pair of temporal nodes which are forbidden
}
\examples{
DBN <- generate_dbn_random_structure(c("A","B","C"), TRUE, 0.5)
DBN_fitted <- generate_dbn_nodes_distributions(DBN,c("A","B","C"), TRUE, 2)
data <- dbn_sampling(DBN_fitted, 2000, 5)
forbidden_edges <- blacklist.temp.DBN(data)
}
