% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_structure_learning.R
\name{blacklist_from_DBN}
\alias{blacklist_from_DBN}
\title{A function returning the blacklist given a DBN object}
\usage{
blacklist_from_DBN(DBN)
}
\arguments{
\item{DBN}{object of class 'DBN'}
}
\value{
data.frame of the arcs in the blacklist
}
\description{
A function returning the blacklist given a DBN object
}
