% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{dbn_forecasting}
\alias{dbn_forecasting}
\title{Generate a sampling dataset}
\usage{
dbn_forecasting(fitted_DBN, observations, timepoints)
}
\arguments{
\item{fitted_DBN}{an object of class 'DBN'}

\item{observations}{a list containing the observations for each variable}

\item{timepoints}{number of timepoints to be forecasted}
}
\value{
a dataframe containing the forecasting
}
\description{
Generate a sampling dataset
}
\examples{
dbn_forecasting(fitted_DBN, observations, timepoints)
}
