% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_parameter_learning.R
\name{DBN_parameters}
\alias{DBN_parameters}
\title{Function for parameter learning in Dynamic Bayesian Networks}
\usage{
DBN_parameters(
  DBN,
  CPTs = list(),
  data = data.frame(),
  replace.unidentifiable = FALSE
)
}
\arguments{
\item{DBN}{object of class 'DBN'}

\item{CPTs}{list of multi-dimensional vector (CPTs for each DBN node)}

\item{data}{data.frame object}

\item{replace.unidentifiable}{If TRUE conditional probabilities for unobserved parents combinations (unidentifiable parameters) are replaced by uniform conditional probabilities, if FALSE (default) they are set as NA}
}
\value{
object of class 'DBN.fit'
}
\description{
Function for parameter learning in Dynamic Bayesian Networks
}
\examples{
learned_dbn <- DBN_parameters(DBN = DBN_example, data = sampling_set)
}
