% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{dbn_sampling}
\alias{dbn_sampling}
\title{Generate a sampling dataset}
\usage{
dbn_sampling(fitted_DBN, N_samples, Time)
}
\arguments{
\item{fitted_DBN}{an object of class 'DBN'}

\item{N_samples}{number of samples}

\item{Time}{time series length}
}
\value{
the generated dataframe
}
\description{
Generate a sampling dataset
}
\examples{
dbn_sampling(DBN_example, N_samples, Time)
}
