% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_initialization.R
\name{define_CPT}
\alias{define_CPT}
\title{Function for Conditionally Probability Table definition by terminal}
\usage{
define_CPT(
  var_name = " ",
  DBN = empty_DBN(dynamic_nodes = c(), markov_order = 1),
  nodes_level = list()
)
}
\arguments{
\item{var_name}{target variable to be conditioned (optional - default: ' ')}

\item{DBN}{object of class 'DBN' (optional - default: an empty DBN with no nodes)}

\item{nodes_level}{named list with the listed levels for each node}
}
\value{
multi-dimensional numerical vector
}
\description{
Function for Conditionally Probability Table definition by terminal
}
\examples{
CPT <- define_CPT() #no information about the variables
B_0.prob <- define_CPT(var_name='B_0', DBN=DBN_example) #definition of the CPT for target variable B_0 given the DBN
}
