% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_initialization.R
\name{define_CPTs}
\alias{define_CPTs}
\title{Function for Conditionally Probability Tables definition by terminal given the Dynamic Bayesian Network}
\usage{
define_CPTs(DBN = empty_DBN(dynamic_nodes = c(), markov_order = 1))
}
\arguments{
\item{DBN}{object of class 'DBN'}
}
\value{
list of multi-dimensional vector (CPTs for each DBN node)
}
\description{
Function for Conditionally Probability Tables definition by terminal given the Dynamic Bayesian Network
}
\examples{
define_CPTs(DBN_example)
}
