% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{dynamic_ordering}
\alias{dynamic_ordering}
\title{Gets the topological ordered nodes list}
\usage{
dynamic_ordering(dbn)
}
\arguments{
\item{structure}{the structure of the network.}
}
\value{
the ordered nodes at time t
}
\description{
This method gets the structure of a DBN, isolates the nodes of the
time slice t and then gives a topological ordering of them.
}
