% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_random.R
\name{fit_random_dbn}
\alias{fit_random_dbn}
\title{fit_random_dbn}
\usage{
fit_random_dbn(
  generated_dbn,
  fixed_cardinality = FALSE,
  max_variables_cardinality
)
}
\arguments{
\item{generated_dbn}{a DBN class object}

\item{fixed_cardinality}{A logical parameter (\code{TRUE} or \code{FALSE}). If \code{TRUE},
all nodes will have the same cardinality as specified by
\code{max_variables_cardinality}. If \code{FALSE}, the cardinality for each node is randomly
determined, with a minimum of 2 and a maximum defined by \code{max_variables_cardinality}.}

\item{max_variables_cardinality}{An integer specifying the maximum cardinality to be used for the nodes.
If \code{fixed_cardinality} is \code{TRUE}, this parameter sets the exact cardinality for all nodes. If \code{fixed_cardinality}
is \code{FALSE}, this parameter defines the upper limit of the range from which the cardinality is randomly selected.}
}
\value{
a DBN.fit object
}
\description{
fit_random_dbn
}
\examples{
fit_random_dbn(generated_dbn, TRUE, 2)
}
