% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_initialization.R
\name{from_DBN_to_G_0}
\alias{from_DBN_to_G_0}
\title{Transformation of a DBN in a G_0 network}
\usage{
from_DBN_to_G_0(DBN)
}
\arguments{
\item{DBN}{object of class 'DBN'}
}
\value{
object of class 'bn'
}
\description{
Transformation of a DBN in a G_0 network
}
\examples{
G_0 <- from_DBN_to_G_0(DBN_example)
}
