% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_initialization.R
\name{from_DBN_to_G_transition}
\alias{from_DBN_to_G_transition}
\title{Transformation of a DBN in a G_transition network}
\usage{
from_DBN_to_G_transition(DBN)
}
\arguments{
\item{DBN}{object of class 'DBN'}
}
\value{
object of class 'bn'
}
\description{
Transformation of a DBN in a G_transition network
}
\examples{
G_transition <- from_DBN_to_G_transition(DBN_example)
}
