% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_parameter_learning.R
\name{from_fitted_DBN_to_fitted_G_transition}
\alias{from_fitted_DBN_to_fitted_G_transition}
\title{Function for G_transition parameters set extraction}
\usage{
from_fitted_DBN_to_fitted_G_transition(DBN_fitted)
}
\arguments{
\item{DBN_fitted}{object of class 'dbn.fit'}
}
\value{
object of class 'bn.fit'
}
\description{
Function for G_transition parameters set extraction
}
\examples{
fitted_transition <- from_fitted_DBN_to_fitted_G_transition(fitted_DBN)
}
