% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_random.R
\name{generate_dbn_nodes_distributions}
\alias{generate_dbn_nodes_distributions}
\title{generate_dbn_nodes_distributions}
\usage{
generate_dbn_nodes_distributions(
  generated_dbn,
  fixed_cardinality,
  max_variables_cardinality
)
}
\arguments{
\item{generated_dbn}{an object of class DBN}

\item{fixed_cardinality}{A logical parameter (\code{TRUE} or \code{FALSE}). If \code{TRUE},
all nodes will have the same cardinality as specified by
\code{max_variables_cardinality}. If \code{FALSE}, the cardinality for each node is randomly
determined, with a minimum of 2 and a maximum defined by \code{max_variables_cardinality}.}

\item{max_variables_cardinality}{An integer specifying the maximum cardinality to be used for the nodes.
If \code{fixed_cardinality} is \code{TRUE}, this parameter sets the exact cardinality for all nodes. If \code{fixed_cardinality}
is \code{FALSE}, this parameter defines the upper limit of the range from which the cardinality is randomly selected.}
}
\value{
a DBN.fit object
}
\description{
This function generates a set of multinomial distributions for each node in a network,
using the Dirichlet distribution to ensure uniform sampling of probability values.
The generated distributions can either have a fixed cardinality for all nodes or a
variable cardinality for each node, depending on the specified parameters.
}
\examples{
generate_dbn_nodes_distributions(generated_dbn,c("A","B","C"), TRUE, 2)
}
