% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_random.R
\name{generate_dbn_random_structure}
\alias{generate_dbn_random_structure}
\title{generate_dbn_random_structure}
\usage{
generate_dbn_random_structure(
  nodes_names,
  is_same,
  prob_edge_intraslice,
  prob_edge_interslice
)
}
\arguments{
\item{nodes_names}{a vector containing the names of nodes.}

\item{is_same}{a boolean, if TRUE g_0 and g_transition will have the same
edges except for the temporal ones}

\item{prob_edge_intraslice}{a double parameter representing the probability of an edge to be present in the intraslice.}

\item{prob_edge_interslice}{a double parameter representing the probability of an edge to be present in the interslice.}
}
\value{
an object of class 'DBN'
}
\description{
Generates a random structure of a DBN. This function uses random.graph
from the bnlearn R package. Generates a DBN structure consistent with the
ordering of the nodes provided by the user; in each arc, the node at the
tail precedes the node at the head in the node set.
Arcs are samples independently with a probability of inclusion specified by
the prob_edge argument. For more details please check:
https://www.bnlearn.com/examples/dag/.
}
\examples{
generate_dbn_random_structure(c("A","B","C"), TRUE, 0.5, 0.4)
}
