% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{get_generic_node_name_0}
\alias{get_generic_node_name_0}
\title{Get the generic name of a node belonging to the graph G_0}
\usage{
get_generic_node_name_0(n)
}
\arguments{
\item{n}{name of node n}
}
\value{
name of node with no time index.
}
\description{
Get the generic name of a node belonging to the graph G_0
}
\examples{
get_generic_node_name_0("A_0")
}
