% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{get_generic_node_name_rex}
\alias{get_generic_node_name_rex}
\title{Get the generic name of a node belonging to the transition graph}
\usage{
get_generic_node_name_rex(n)
}
\arguments{
\item{n}{name of node n}
}
\value{
name of node with no time index.
}
\description{
Get the generic name of a node belonging to the transition graph
}
\examples{
get_generic_node_name_rex("A_t-1")
}
