% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{get_index_regular_expression}
\alias{get_index_regular_expression}
\title{Get the time index of a node}
\usage{
get_index_regular_expression(n)
}
\arguments{
\item{n}{name of node n}
}
\value{
time index of a node n.
}
\description{
Get the time index of a node
}
\examples{
get_index_regular_expression("A_t-1")
}
