% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_random.R
\name{get_node_name_vector}
\alias{get_node_name_vector}
\title{get_node_name_vector}
\usage{
get_node_name_vector(node_name)
}
\arguments{
\item{node_name}{a string representing the name of a node}
}
\value{
a vector representing the node name.
}
\description{
the function transforms the node name to the vector format A_t-1 -> c("A", "t-1")
}
\examples{
get_node_name_vector("A_0")
}
