% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{get_nodes_t}
\alias{get_nodes_t}
\title{Return the variables indexes by t in G_transition}
\usage{
get_nodes_t(G_transition)
}
\arguments{
\item{G_transition}{a G_transition graph.}
}
\value{
nodes indexed by t only e.g. A_t, B_t.
}
\description{
Return the variables indexes by t in G_transition
}
\examples{
get_nodes_t(G_transition)
}
