% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{get_parents}
\alias{get_parents}
\title{Get the parents of a node}
\usage{
get_parents(G_transition, n)
}
\arguments{
\item{n}{name of node n}

\item{G}{A bn.fit object}
}
\value{
parents of the node n.
}
\description{
Get the parents of a node
}
\examples{
get_parents(G, "A_t")
}
