% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_sampling.R
\name{get_time_point}
\alias{get_time_point}
\title{Get the time point to retrieve the values of variables in the time series}
\usage{
get_time_point(t, index_n)
}
\arguments{
\item{t}{time point}

\item{index_n}{time index of node n}
}
\value{
time-point in the series used to retrieve the value of the node n
}
\description{
Get the time point to retrieve the values of variables in the time series
}
\examples{
get_time_point(3,1)
}
