% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_unrolled_dbn}
\alias{get_unrolled_dbn}
\title{get_unrolled_dbn}
\usage{
get_unrolled_dbn(dbn_fitted, slices)
}
\arguments{
\item{dbn_fitted}{a dbn.fit class object}

\item{slices}{number of time-slice of the unrolled network}
}
\value{
a bn.fit object
}
\description{
this function generates an unrolled dynamic bayesian network
}
\examples{
get_unrolled_dbn(my_fitted_dbn, 4)
}
