% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{node_levels_g_0}
\alias{node_levels_g_0}
\title{Defines a level for every node in the network g_0}
\usage{
node_levels_g_0(net, order, lvl = 1, acc = NULL)
}
\arguments{
\item{net}{the object of class bn G_0}

\item{order}{a topological order of the nodes, with the orphan nodes
in the first place. See \code{\link{node.ordering}}}

\item{lvl}{current level being processed}

\item{acc}{accumulator of the nodes already processed}
}
\value{
a matrix with the names of the nodes in the first row and their
level in the second
}
\description{
Calculates the levels in which the nodes will be distributed when plotting
the structure. This level is defined by their parent nodes: a node with no
parents will always be in the level 0. Subsequently, the level of a node
will be one more of the maximum level of his parents.
}
