% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_structure_learning.R
\name{pc.stable.DBN}
\alias{pc.stable.DBN}
\title{Learn the equivalence class of DBN via PC-stable algorithm}
\usage{
pc.stable.DBN(
  data,
  test = "mi",
  max.sx = NULL,
  blacklist = NULL,
  whitelist = NULL,
  ...
)
}
\arguments{
\item{data}{a data.frame to be given as an input the to structure learning algorithm}

\item{test}{a character string, the label of the conditional independence test to be used in the algorithm test (default is \verb{Mutual Information} mi)}

\item{max.sx}{an integer, the maximum number of parents for each node (default is no limit)}

\item{blacklist}{a matrix with columns 'from' and 'to' defining the set of arcs that are forbidden}

\item{whitelist}{a matrix with columns 'from' and 'to' defining the set of arcs that are fixed}

\item{...}{additional parameters for the chosen test}
}
\value{
An object of class 'DBN', which is a DBN learned via PC stable algorithm
}
\description{
Learn the equivalence class of DBN via PC-stable algorithm
}
\examples{
DBN <- generate_dbn_random_structure(c("A","B","C"), TRUE, 0.5)
DBN_fitted <- generate_dbn_nodes_distributions(DBN,c("A","B","C"), TRUE, 2)
data <- dbn_sampling(DBN_fitted, 2000, 5)
pc.stable.DBN(data)

}
