% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{plot_g_transition}
\alias{plot_g_transition}
\title{Plots the network G_transition}
\usage{
plot_g_transition(dbn)
}
\arguments{
\item{dbn}{a list representing a dbn}
}
\description{
Plots the network G_transition
}
