% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_nodes_unroll}
\alias{rename_nodes_unroll}
\title{rename_nodes_unroll}
\usage{
rename_nodes_unroll(time_slice, nodes_names)
}
\arguments{
\item{time_slice}{an integer representing a time slice}

\item{nodes_names}{a list of nodes names}
}
\description{
Renames a list of nodes, making nodes names compliant to the dbn unrolled structure
}
