% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{select_nodes_names_with_t}
\alias{select_nodes_names_with_t}
\title{Filters out all the nodes names that do not end up with t}
\usage{
select_nodes_names_with_t(nodes_list)
}
\arguments{
\item{nodes_list}{a list containing nodes names}
}
\value{
a filtered list with nodes names that end up with t
}
\description{
Filters out all the nodes names that do not end up with t
}
