% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_visualization.R
\name{select_nodes_names_with_t_minus_n}
\alias{select_nodes_names_with_t_minus_n}
\title{Filters out all the nodes names that do not end up with t-n}
\usage{
select_nodes_names_with_t_minus_n(nodes_list, n)
}
\arguments{
\item{nodes_list}{a list containing nodes names}

\item{n}{time point}
}
\description{
Filters out all the nodes names that do not end up with t-n
}
