% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_structure_learning.R
\name{StructureLearning.Data}
\alias{StructureLearning.Data}
\title{Generation of a Dynamic Bayesian Network given two compatible networks \eqn{G_0} and \eqn{G_{transition}}}
\usage{
StructureLearning.Data(PN, TN, nodes)
}
\arguments{
\item{PN}{\eqn{G_0}, an object of class 'bn'}

\item{TN}{\eqn{G_{transition}}, an object of class 'bn'}

\item{nodes}{character vector of node names}
}
\value{
An object of class 'DBN' with the characteristics of \eqn{G_0} and \eqn{G_{transition}}
}
\description{
Generation of a Dynamic Bayesian Network given two compatible networks \eqn{G_0} and \eqn{G_{transition}}
}
\examples{
DBN <- generate_dbn_random_structure(c("A","B","C"), TRUE, 0.5)
DBN_fitted <- generate_dbn_nodes_distributions(DBN,c("A","B","C"), TRUE, 2)
data <- dbn_sampling(DBN_fitted, 2000, 5)
c(data0, dataTR, ., blacklistTR, ., .) \%<-\% StructureLearning.Preprocess(data = data)
bn_sampled_0 <- pc.stable(data0)
bn_sampled_TR <- pc.stable(dataTR, blacklist = blacklistTR)
StructureLearning.Data(PN = bn_sampled_0, TN =  bn_sampled_TR, nodes = setdiff(names(data), c('Sample_id','Time')))

}
