% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbn_structure_learning.R
\name{whitelist_from_DBN}
\alias{whitelist_from_DBN}
\title{A function returning the whitelist given a DBN object}
\usage{
whitelist_from_DBN(DBN)
}
\arguments{
\item{DBN}{object of class 'DBN'}
}
\value{
data.frame of the arcs in the whitelist
}
\description{
A function returning the whitelist given a DBN object
}
